# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Podcast API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/episodes": {
            "get": {
                "tags": [
                    "Episodes"
                ],
                "summary": "Liste tous les épisodes",
                "operationId": "39bb119c426c59c593560bf467c00fe9",
                "responses": {
                    "200": {
                        "description": "Liste des épisodes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Episodes"
                ],
                "summary": "Créer un nouvel épisode",
                "operationId": "1977d240cbbdd586e14ae1cda09bdb61",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "podcast_id",
                                    "title",
                                    "audio_url"
                                ],
                                "properties": {
                                    "podcast_id": {
                                        "type": "integer"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "audio_url": {
                                        "type": "string"
                                    },
                                    "duration": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Épisode créé"
                    }
                }
            }
        },
        "/api/episodes/{id}": {
            "get": {
                "tags": [
                    "Episodes"
                ],
                "summary": "Afficher un épisode",
                "operationId": "682c9f0b582614c5d25814ad0a6659c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'épisode",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails d'un épisode"
                    }
                }
            },
            "put": {
                "tags": [
                    "Episodes"
                ],
                "summary": "Modifier un épisode",
                "operationId": "7c242aee03c0c562d61e184cc40ffbcf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Épisode mis à jour"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Episodes"
                ],
                "summary": "Supprimer un épisode",
                "operationId": "2b31a0a0e3d3c3f435a21fe63779b608",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Épisode supprimé"
                    }
                }
            }
        },
        "/api/hosts": {
            "get": {
                "tags": [
                    "Hosts"
                ],
                "summary": "Lister tous les hosts",
                "operationId": "830cd25d337656d5bc42faa7b8565db5",
                "responses": {
                    "200": {
                        "description": "Liste des hosts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Host"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Hosts"
                ],
                "summary": "Créer un nouveau host",
                "operationId": "87ac5d8eb9c7da60122f373f261566b3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Host Example"
                                    },
                                    "bio": {
                                        "type": "string",
                                        "example": "Biographie du host"
                                    },
                                    "avatar": {
                                        "type": "string",
                                        "example": "avatar.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Host créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Host"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/hosts/{id}": {
            "get": {
                "tags": [
                    "Hosts"
                ],
                "summary": "Afficher un host par ID",
                "operationId": "b3fd9e7fb77c05d2c56d9d986af84416",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du host",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Host trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Host"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Host non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Hosts"
                ],
                "summary": "Mettre à jour un host",
                "operationId": "3a9629ba010ad7dd37bbf39c531f4d54",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du host",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "bio": {
                                        "type": "string"
                                    },
                                    "avatar": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Host mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Host"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Hosts"
                ],
                "summary": "Supprimer un host",
                "operationId": "d758087e4042672a7896df7123cb105c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du host",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Host supprimé"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Host": {
                "title": "Host",
                "description": "Modèle Host",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "bio": {
                        "type": "string",
                        "example": "Podcaster et auteur"
                    },
                    "avatar": {
                        "type": "string",
                        "example": "avatar.jpg"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01 12:00:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01 12:00:00"
                    }
                },
                "type": "object"
            }
        }
    },
...
}